+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Unreleased]
 - Start tracking changes in `CHANGELOG.md`
+- Aggregate over connection types in the cred explorer (#502)